python/requirements.txt
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.0.1
onnxruntime==1.18.0
pdfminer.six==20231228
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
rapidfuzz==3.9.3
reportlab==4.2.2
protobuf==5.28.0
typer==0.12.5
